% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giftR-package.R
\docType{package}
\name{giftR}
\alias{giftR-package}
\alias{giftR}
\title{giftR: A White Elephant Gift Exchange Helper}
\description{
The giftR package provides tools to facilitate a white elephant gift exchange.
}
\author{
\strong{Maintainer}: Carson Slater \email{Carson_Slater1@baylor.edu}

}
