% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GIFTrData}
\alias{GIFTrData}
\title{DATASET_TITLE}
\format{A data frame with 10 rows and 9 variables:
\describe{
  \item{\code{categories}}{character COLUMN_DESCRIPTION}
  \item{\code{names}}{character COLUMN_DESCRIPTION}
  \item{\code{question}}{character COLUMN_DESCRIPTION}
  \item{\code{A}}{character COLUMN_DESCRIPTION}
  \item{\code{B}}{character COLUMN_DESCRIPTION}
  \item{\code{C}}{character COLUMN_DESCRIPTION}
  \item{\code{D}}{character COLUMN_DESCRIPTION}
  \item{\code{E}}{character COLUMN_DESCRIPTION}
  \item{\code{question_type}}{character COLUMN_DESCRIPTION}
}}
\usage{
data(GIFTrData)
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
