% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_q.R
\name{num_q}
\alias{num_q}
\title{Create GIFT File with Numeric Entery Questions From Spreadsheet}
\usage{
num_q(data, questions, answers, categories, question_names = NULL,
  output)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{questions}{PARAM_DESCRIPTION}

\item{answers}{PARAM_DESCRIPTION}

\item{categories}{PARAM_DESCRIPTION, Default: NULL}

\item{question_names}{PARAM_DESCRIPTION, Default: NULL}

\item{output}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[GIFTr]{GIFTr}}
}
