% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santa.R
\name{santa}
\alias{santa}
\title{secret santa function}
\usage{
santa(p.df, email = FALSE, organizer = NULL, subject = NULL,
  host.nm = "smtp.gmail.com", prt = 465, usr.nm = NULL, pswd = NULL,
  giftLim = 20)
}
\arguments{
\item{p.df}{Data frame with information for each individual. Columns need
to be 'giver' and 'couple', with 'email' optional. The 'couple' identifies
significant others numerically (i.e., spouses have the same number)}

\item{email}{Logical argument for whether or not participants should
automatically be emailed; defaults to FALSE}

\item{organizer}{If email=TRUE, email address of the organizer}

\item{subject}{If email=TRUE, subject of the email}

\item{host.nm}{If email=TRUE, host.name for mailR; defaults to "smtp.gmail.com"}

\item{prt}{If email=TRUE, smtp port; defaults to 465}

\item{usr.nm}{If email=TRUE, email username}

\item{pswd}{If email=TRUE, email/application email}

\item{giftLim}{If email=TRUE, the gift limit entered just as a number}
}
\description{
This function generates random assignments or secret santa and emails the
participants if desired.
}
\note{
Sending emails requires installation of the mailR package. If you are using
Gmail, the two-factor authentication requires an application password - not your
normal account password.
}
\examples{
# A group with two couples and two additional individuals
p.df <- data.frame(giver=LETTERS[1:6],
                   couple=c(1,1,2,2,3,4))
santa(p.df, email=FALSE)

p.df$email <- "tisz1545@colorado.edu"
santa(p.df, email=TRUE, organizer="tim.szewczyk@gmail.com",
      subject="2017 Secret Santa", usr.nm="tim.szewczyk",
      pswd="xxxxxxxxxxxx", giftLim=20)
}
\keyword{santa}
\keyword{secret}
