% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_ans.r
\name{short_ans}
\alias{short_ans}
\title{Create GIFT File with Short Answer Questions From Spreadsheet}
\usage{
short_ans(data, questions, answers, categories, question_names = NULL,
  output)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{questions}{PARAM_DESCRIPTION}

\item{answers}{PARAM_DESCRIPTION}

\item{categories}{PARAM_DESCRIPTION, Default: NULL}

\item{question_names}{PARAM_DESCRIPTION, Default: NULL}

\item{output}{PARAM_DESCRIPTION}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[GIFTr]{GIFTr}}
}
